Initial Commit.
From talking on the harp IRC channel, decided to make a WordPress to
JSON conversion tool.

First commit, so it's simple. Pulls just the posts and their meta data
out of the database, binds it to an object, then spits out the object as
a JSON. In a list that could be parsed.

You can pipe the JSON to a file by running the script and directing it
to a file, or you could do whatever floats your boat.