json escaping of object keys for _data
Because we're using the titles of the posts as the keys into the json
data, quotes inside of them cause syntax issues. SO, we escape them by
calling json dumps on it.

Also, added flag to kill off non ascii characters. This is really me
messing with a database that has some funky encoding and I got tired of
dealing with not being able to pipe out to a file without jumping
through hoops.

Added a stripping regex/function so strip out data before it gets
encoded, this can be used to spit everything out onto one line.