First pass as unserializing objects from WP.
Using phpserialize to load the objects, the object hook should be
mapping fields, but it throws ValueErrors when it can't load one for
some reason. This seems to happen ofen and I haven't figured out if the
data I'm using is simply corrupted in some way, or if it's failing for
all  unserilaizations.